web: Move everything in a package
