+++ web/app/templates/graphic_widget.html
