+++ web/app/templates/lateral_zone_menu.html
