+++ web/app/templates/stats_widget.html
