+++ web/app/templates/zone_summary_widget.html
